refactor(codebase): Update architecture
Remove all dead code from previous tests to keep only `dlsp` official codebase.
