+++ README.md
@@ -1,3 +1,14 @@
-# TODO
+# Dagger LSP
 
-First, only consider the `small-repro` folder. `go get`, then `go build -o dlsp`. Find a way to add this binary to your PATH (the vscode extension basically just runs the `dlsp` command to connect with it)
+Language Server implementation for Dagger.
+
+### Build
+
+```shell
+go build -o dlsp
+```
+
+### Capabilities
+
+In progress...
+
