--- lsp/util.go
@@ -1,10 +1,0 @@
-package lsp
-
-func boolPtr(v bool) *bool {
-	b := v
-	return &b
-}
-
-func isTrue(v *bool) bool {
-	return v != nil && *v == true
-}
