+++ plan/kind_string.go
